Do not throw excpetion when testData directory does not exist.

+++ core/src/main/java/org/openda/blackbox/config/BBUtils.java
@@ -67,6 +67,7 @@ public class BBUtils {
     }
 
 	public static void makeDirectoryCloneForTestRun(File source, File target) {
+		if (source==null) return;
 		checkSourceAndTarget(source, target, true);
 		if (target.exists()) {
 			deleteDirectory(target);
